move routespec to its own file
